test to check validator count. cleanup test utlis
