+++ ol/genesis-tools/Cargo.toml
@@ -39,4 +39,4 @@ reqwest = { version = "0.11.2", features = ["blocking", "json"], default_feature
 tokio-test = "*"
 diem-genesis-tool = { path = "../../config/management/genesis", version = "0.1.0" }
 diem-temppath = { path = "../../crates/diem-temppath", version = "0.1.0" }
-
+# diemdb = { path = "../../storage/diemdb", version = "0.1.0" }
